requirements.txt
flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
plotly==5.17.0
weasyprint==60.2
markdown==3.5.2
werkzeug==3.0.1
lxml==4.9.3
beautifulsoup4==4.12.2
pandas==2.1.4
reportlab==4.0.7
